+++ whale-line-iconify.el
@@ -40,20 +40,24 @@ This has no effect if icons cannot be enabled. See
 (defcustom whale-line-iconify-specs
   '((project . (:name "package" :font octicon :face whale-line-emphasis))
     (vc . (:name "code-fork" :face whale-line-contrast))
-    (buffer-read-only . (:name "lock" :face whale-line-contrast :fallback "@" :parent buffer-status))
-    (buffer-file-name . (:name "sticky-note-o" :face whale-line-shadow :fallback "&" :parent buffer-status))
-    (buffer-modified . (:name "pencil" :face whale-line-emphasis :fallback "*" :parent buffer-status))
-    (window-dedicated . (:name "link" :face whale-line-shadow :fallback "^" :parent window-status))
-    (window-no-other . (:name "low-vision" :face whale-line-shadow :fallback "~" :parent window-status))
+    (major-mode . (:function all-the-icons-icon-for-buffer))
+    (buffer-read-only . (:name "lock" :face whale-line-contrast :parent buffer-status))
+    (buffer-file-name . (:name "sticky-note-o" :face whale-line-shadow :parent buffer-status))
+    (buffer-modified . (:name "pencil" :face whale-line-emphasis :parent buffer-status))
+    (window-dedicated . (:name "link" :face whale-line-shadow :parent window-status))
+    (window-no-other . (:name "low-vision" :face whale-line-shadow :parent window-status))
     (buffer-fallback . (:name "question-circle" :face whale-line-contrast :no-defaults t))
-    (lsp . (:name "plug" :face whale-line-contrast :fallback "LSP"))
-    (debug . (:name "bug" :face whale-line-urgent :fallback "BUG"))
-    (partial-recall . (:name "tag" :face whale-line-contrast :fallback "PR")))
+    (lsp . (:name "plug" :face whale-line-contrast))
+    (debug . (:name "bug" :face whale-line-urgent)))
   "Icon specifications.
 
 This is an alist of (ICON-SYMBOL . SPECS). The ICON-SYMBOL is the
 symbol passed to `whale-line-iconify' to retrieve the associated
 specs. SPECS is a plist with the following keys.
+
+FUNCTION is a function to use instead of constructing it using
+NAME. If PASS-ARGS is t, all icon specs are passed to the
+function.
 
 NAME is the name of the icon in the font. The remaining keys are
 optional.
@@ -62,9 +66,6 @@ FACE is the face to use for the icon.
 
 FONT is a symbol, namely one of the possible suffixes of
 `all-the-icons-insert-*'. It defaults to `faicon'.
-
-FALLBACK is a string to use when the icon can't be used or is
-disabled.
 
 PARENT is a symbol of another icon. This allows disabling an icon
 through its parent.
@@ -86,18 +87,21 @@ All remaining keys and their values will be passed to
 
 (defun whale-line-iconify--from-specs (specs)
   "Get icon from SPECS."
-  (let* ((font (or (plist-get specs :font) 'faicon))
-         (fun (intern (concat "all-the-icons-" (symbol-name font))))
-         (icon (plist-get specs :name))
-         (remainder (whale-line-iconify--pure-specs specs))
-         (defaults (whale-line-iconify--default-specs specs remainder)))
+  (if-let* ((fun (plist-get specs :function)))
+      (apply fun (when (plist-get specs :pass-args)
+                   (whale-line-iconify--pure-specs specs)))
+    (let* ((font (or (plist-get specs :font) 'faicon))
+           (fun (intern (concat "all-the-icons-" (symbol-name font))))
+           (icon (plist-get specs :name))
+           (remainder (whale-line-iconify--pure-specs specs))
+           (defaults (whale-line-iconify--default-specs specs remainder)))
 
-    (apply fun (append (list icon) remainder defaults))))
+      (apply fun (append (list icon) remainder defaults)))))
 
 (defun whale-line-iconify--pure-specs (specs)
   "Return icon only specs from SPECS."
   (cl-loop for (key . val) in (cl--plist-to-alist specs)
-           unless (memq key '(:name :fallback :font :no-defaults :parent))
+           unless (memq key '(:name :function :font :no-defaults :parent :pass-args))
            nconc (list key val)))
 
 (defun whale-line-iconify--default-specs (specs existing)
@@ -129,20 +133,13 @@ Only properties not in EXISTING are added."
   "Get the icon for NAME.
 
 If optional argument FACE is passed, it will be used instead of
-the value in the retrieved specs.
+the value in the retrieved specs."
+  (when-let* ((specs (copy-tree (cdr-safe (assoc name whale-line-iconify-specs))))
+              (specs (if face (plist-put specs :face face) specs))
+              (parent (or (plist-get specs :parent) name))
+              ((whale-line-iconify--use-for-p parent)))
 
-If icon can't or shouldn't be displayed, any existing fallback is
-returned."
-  (if-let* ((specs (copy-tree (cdr-safe (assoc name whale-line-iconify-specs))))
-            (specs (if face (plist-put specs :face face) specs))
-            (parent (or (plist-get specs :parent) name))
-            ((whale-line-iconify--use-for-p parent)))
-
-      (whale-line-iconify--from-specs specs)
-
-    (when-let ((fallback (plist-get specs :fallback)))
-
-      (propertize fallback 'face (plist-get specs :face)))))
+    (whale-line-iconify--from-specs specs)))
 
 (provide 'whale-line-iconify)